requirements.txt
requests
python-dateutil
google-generativeai
PyYAML
PyGithub
numpy
